Prevent piece capture on illegal move
You could capture a piece on a move that was shorter/longer than a valid
placement.

For example: if you rolled a 2 and there was an enemy piece 1 space
away. You could place your token in space 1, remove the enemy piece, but
your move wouldn't count as valid and you'd snap back to your original
position. By placing the catpure method after the "in_place" we ensure
we only remove the enemy piece if its fully valid.